Добавил "Глеба-пидора", эт оч важная фича!

+++ src/main/java/com/ecs/measure/GUI/Color.java
@@ -9,9 +9,9 @@ public class Color {
         BLUE = new Color(0, 0, 1, 1),
         TRANSPARENT = new Color(0, 0, 0, 0);
     
-    public float r, g, b, a;
+    public double r, g, b, a;
 
-    public Color(float r, float g, float b, float a) {
+    public Color(double r, double g, double b, double a) {
         this.r = r;
         this.g = g;
         this.b = b;
@@ -19,17 +19,17 @@ public class Color {
     }
 
     public Color(int color) {
-        this.a = color > 0xFFFFFF ? (float) (color >> 24 & 0xFF) / 255 : 1;
-        this.r = (float) (color >> 16 & 0xFF) / 255;
-        this.g = (float) (color >> 8 & 0xFF) / 255;
-        this.b = (float) (color & 0xFF) / 255;
+        this.a = color > 0xFFFFFF ? (double) (color >> 24 & 0xFF) / 255 : 1;
+        this.r = (double) (color >> 16 & 0xFF) / 255;
+        this.g = (double) (color >> 8 & 0xFF) / 255;
+        this.b = (double) (color & 0xFF) / 255;
     }
 
     public static int toARGB(Color color) {
         return ((int) (color.a * 255) << 24) | ((int) (color.r * 255) << 16) | ((int) (color.g * 255) << 8) | ((int) (color.b * 255));
     }
     
-    public static Color transition(Color color1, Color color2, float factor) {
+    public static Color transition(Color color1, Color color2, double factor) {
         return new Color(
             color1.r + (color2.r - color1.r) * factor,
             color1.g + (color2.g - color1.g) * factor,
